Extract code into more focused files
